Part 1: Removal of various warnings supression and fixing their warnings, chagning class and delete C++ keywords in names, and changes related to compile common target as C++

+++ src/char/geoip.c
@@ -158,7 +158,7 @@ static void geoip_init(void)
 		return;
 	}
 	geoip->data->cache = aMalloc(sizeof(unsigned char) * bufa.st_size);
-	if (fread(geoip->data->cache, sizeof(unsigned char), bufa.st_size, db) != bufa.st_size) {
+	if (fread(geoip->data->cache, sizeof(unsigned char), bufa.st_size, db) != (size_t)bufa.st_size) {
 		ShowError("geoip_cache: Couldn't read all elements!\n");
 		fclose(db);
 		geoip->final(false);
