+++ src/common/memmgr.c
@@ -620,7 +620,7 @@ static void mfree_(void *ptr, const char *file, int line, const char *func)
 	} else {
 		/* Release unit */
 		struct block *block = head->block;
-		if( (char*)head - (char*)block > sizeof(struct block) ) {
+		if( (char*)head - (char*)block > (ssize_t)sizeof(struct block) ) {
 			ShowError("Memory manager: args of aFree 0x%p is invalid pointer %s line %d\n", ptr, file, line);
 		} else if(head->block == NULL) {
 			ShowError("Memory manager: args of aFree 0x%p is freed pointer %s:%d@%s\n", ptr, file, line, func);
